requirements.txt
 # Retail Intelligence Platform Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Data Processing
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2

# Machine Learning & Forecasting
prophet==1.1.4
xgboost==2.0.1
statsmodels==0.14.0
scikit-learn==1.3.2

# Visualization & Reporting
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Database & Storage
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
jinja2==3.1.2

# Indian Currency & Localization
babel==2.13.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
